+++ docs/RELEASE-NOTES.rst
@@ -9,6 +9,7 @@ Added Functionality
 **What's new:**
     * Multi Cluster
     * CRD
+    * Support for Calico CNI with Static Routing Mode, See `Documentation <https://github.com/F5Networks/k8s-bigip-ctlr/blob/2.x-master/docs/config_examples/StaticRoute/README.md>`_
 Bug Fixes
 ````````````
 * `Issue 3371 https://github.com/F5Networks/k8s-bigip-ctlr/issues/3371`_: CIS added irules cannot have "event disable all"